time sort for today games, and all changes required
