+++ src/helpers/helperFunctions.tsx
@@ -1,4 +1,4 @@
-import { playerStatsDataType } from '../types/basketballdata';
+import { playerStatsDataType, basketballData } from '../types/basketballdata';
 
 function minutesSort(a: playerStatsDataType, b: playerStatsDataType) {
     return parseInt(b.min) - parseInt(a.min);
@@ -20,4 +20,21 @@ function formatTimeInET(dateString: string) {
     }
 }
 
-export { formatTimeInET, minutesSort };
+function gameStartTimeSort(a: basketballData, b: basketballData) {
+    // a/b object will have a status which will be a string: XX:XX PM EST
+    // Get the time part only XX:XX
+    const timeA = a.status.split(' ')[0];
+    const timeB = b.status.split(' ')[0];
+    // Split around the colons, hours and minutes
+    const [hourA, minuteA] = timeA.split(':').map(Number);
+    const [hourB, minuteB] = timeB.split(':').map(Number);
+    // Convert 24 hour to 12 and then convert all to minutes for sorting.
+    const timeValueA =
+        hourA >= 12 ? (hourA - 12) * 60 + minuteA : hourA * 60 + minuteA;
+    const timeValueB =
+        hourB >= 12 ? (hourB - 12) * 60 + minuteB : hourB * 60 + minuteB;
+
+    return timeValueA - timeValueB;
+}
+
+export { formatTimeInET, gameStartTimeSort, minutesSort };
